Update service and job names
